requirements.txt
torch==2.3.0
torchvision
torchaudio
yolov5
sqlalchemy
pymysql
numpy
opencv-python
pandas
Pillow
jupyter